Add bbclass for recipe-specific sysroots
[YOCTO #1560]

WARNING: This is a work in progress.

General Remarks
===============
This feature is meant to be enabled by putting
  INHERIT += "isolated-sysroots"
in your local.conf. While the idea is to make this optional until it is
ready for use by everybody, I have not yet tried building with these
changes and isolated-sysroots.bbclass disabled, so this probably breaks
normal builds right now.

The simplest approach to recipe-specific sysroots would be a single task
run before do_configure that populates a recipe-specific sysroot.
Unfortunately, that comes with a couple of problems:
 - Tasks before do_configure already have dependencies. For example,
   do_fetch might need git-native.
 - Inside a task, it is currently impossible to find the dependencies of
   other tasks of the same recipe. BB_TASKDEPDATA only contains the
   dependencies of the current task.
 - Dependencies might not be available in the sysroot in a pre-task. For
   example, a dependency of do_build might not be available for copying
   into the recipe specific sysroot during do_configure.

Populating Sysroots
===================
Instead, this approach patches bitbake-worker to populate the sysroot
before a task is run. See the changes in bitbake/bin/bitbake-worker,
which implement this. Keep in mind that multiple bitbake-workers for
different tasks of the same recipe can be run in parallel, so
appropriate locking is required.

The recipe specific sysroot is ${PRIVATE_STAGING_DIR}, which is in
${WORKDIR}. It is populated by hardlinking files from
${ISOLATED_STAGING_DIR}/${DEP_PN} foreach $DEP_PN in the dependencies.
Note that the current implementation is very unoptimized and will always
hardlink all files for all tasks regardless of whether they have already
been copied. Storing the task hashes of the do_populate_sysroot tasks of
each dependency should allow simplifying this a lot (but the purpose of
this implementation was to show that it can be done, not to make it as
fast as normal builds).

Additionally, file conflicts in the sysroot are currently not handled.

There is no sstate cache for the recipe specific sysroot, because
restoring the do_populate_sysroot sstate and re-creating the hardlinks
is probably faster then re-extracting all sysroots from sstate cache.

Changes to do_populate_sysroot
==============================
isolated-sysroots.bbclass introduces a separate location recipes should
write to in order to get new files into the sysroot areas, called
${WRITE_STAGING_DIR}. With isolated sysroots enabled, this defaults to
${ISOLATED_STAGING_DIR}/${PN}, with isolated sysroots disabled, it falls
back to ${STAGING_DIR} in an attempt to work just like shared sysroots.

A change in staging.bbclass deals with most recipes by writing the
output of ${SYSROOT_DESTDIR} to ${WRITE_STAGING_DIR_HOST}. Recipes that
use ${STAGING_DIR} for writing directly (e.g. because they configure
using this path as prefix) need to be fixed.

Special Case: pseudo
====================
The pseudo binary is a special case in this, because it needs to be
available before the recipe specific sysroot is populated. Because the
default path points to a recipe specific sysroot with these changes,
isolated-sysroots.bbclass changes a number of FAKEROOT* variables.

Changes to native.bbclass
=========================
Native tools turn out to be the biggest problem in this change. Before
this change, they are configured with a prefix of ${STAGING_DIR_NATIVE}.
With this change, all native tools are moved to different locations
before they are run, which causes problems for some programs that try to
find resources installed by other recipes relative to their configured
prefix (which they fail to do because the prefix values are now
different per recipe). We considered keeping a shared sysroot for native
tools or configuring all native tools for the same prefix (/usr being
a bad choice, /yocto/native-sysroot/ probably being a better choice),
but we would likely have had to fix more of the recipes to work with
this configuration.

For now, native recipes are configured with a prefix pointing to their
own isolated sysroot area. Some native tools had to be fixed to work in
this configuration.

TODO
====
 - Improve the performance of the critical path; add parallelization
 - Add sanity checking to ensure files installed by -native recipes are
   all in the expected $WRITE_STAGING_DIR prefix (this will be violated
   if prefixes from dependencies leak over)
 - Improve handling of incremental builds and re-use recipe specific
   sysroot content between tasks

Signed-off-by: Clemens Lang <clemens.lang@bmw-carit.de>