isolated sysroots: support pkgdata
[YOCTO #1560]

packagedata used to write into the shared sysroot area. When enabling
isolated-sysroots.bbclass, this shared area no longer exists, so
packagedata should either (for dependencies) be in the recipe specific
sysroot, or (for packages created by the current recipe) in the sysroot
area for this recipe.

Some calls to functions provided by packagedata.bbclass require
information about the packages created by the current recipe, so adjust
the bbclass to search for the current recipe first, and in the recipe
specific sysroot area second. Without this change, packaging fails
because expansion of SUMMARY_${PN}-{dev,doc,...} fails due to a circular
variable reference. Since this is not obvious, I am mentioning it here
in order to avoid long debug sessions should you run into the same
problem.

Signed-off-by: Clemens Lang <clemens.lang@bmw-carit.de>