siteconfig: Support isolated sysroots
do_siteconfig_gencache() uses the compilers (and thus,
$TOOLCHAIN_OPTIONS) to pre-populate presence and location of a couple of
headers and other configure checks. $TOOLCHAIN_OPTIONS contains include
flags for $STAGING_DIR, which does not contain the files provided by the
current recipe when isolated sysroots are enabled.

This would not cause the siteconfig generation to fail, but would
generate a cache with incorrect information, which other recipes would
rely on and then in turn possibly fail to build or build differently.

Fix the issue by temporarily including ${SYSROOT_DESTDIR} in linker and
preprocessor search paths for the execution of do_siteconfig_gencache().

Signed-off-by: Clemens Lang <clemens.lang@bmw-carit.de>