docbook-utils-native: Make relocatable
The jw utility provided by docbook-utils-native and used by openjade
searches for files installed by other packages relative to its own
configured prefix. This fails when isolated sysroots are used.

Apply a patch to determine the prefix at runtime relative to the path of
the binary and use that for searching resources.

Signed-off-by: Clemens Lang <clemens.lang@bmw-carit.de>