flex: Fall back to "m4" in $PATH if not found
Flex searches for the "m4" binary in a path determined at compile time.
This fails if the copy of m4 is in a temporary location such as an
recipe specific sysroot of flex.

Add code to fall back to "m4" from the $PATH environment variable if the
configure-time location is not available.

Signed-off-by: Clemens Lang <clemens.lang@bmw-carit.de>