perl: Fix build with isolated-sysroots
Use the ${WRITE_STAGING_*} variables for outputs written from the perl
build system to support isolated-sysroots.bbclass.

Additionally, pass -Dinstallstyle="lib/perl5" to perl's configure
script, because perl will switch this default to a different value if
the string 'perl' appears in its target prefix (which would have caused
builds to fail already if somebody had every built Yocto in a path that
contained the string 'perl'). With isolated sysroots, the name of the
recipe always appears in the target prefix of native packages, so this
is always triggered.

Modify the perl5-native wrapper scripts to compute the paths to their
library directories at runtime to make perl5-native relocatable as
required by recipe specific sysroots.

Signed-off-by: Clemens Lang <clemens.lang@bmw-carit.de>