sgml-common-native: Make relocatable
The sgmlwhich command installed by sgml-common-native is used to locate
other SGML resources in the sysroot, but defaults to searching a path
passed at configure time, which breaks when recipe specific sysroots are
enabled because the path changes per recipe.

Patch sgmlwhich to search relative to its own location to fix this.

Signed-off-by: Clemens Lang <clemens.lang@bmw-carit.de>