python{,-native}: Fix build with isolated sysroots
Python's build system has a number of problems that cause the build to
fail when recipe specific sysroots are enabled.

First, Python ships a copy of libffi and the recipe runs autoreconf for
this copy, but this autoreconf fails to find resources installed by
automake because it expects those resources in its configured prefix.
With isolated-sysroots.bbclass, automake gets a different installation
prefix, though. autotools.bbclass sets a variable named ${ACLOCAL} to
the path and flags required for aclocal to find those resources, but the
copy of autoreconf in Python's recipe did not use this variable.

Additionally, Python installs a header file in $prefix/include and later
parses this header file again to find a number of configure-time
variables specific to this installation of Python. Because this header
is architecture-specific, the Python recipe post-processes it with
oe_multilib_header, which replaces the file with a #if-guarded #include
of the original file, which is renamed with a architecture suffix. This
breaks Python's parsing of the file, because it does not actually follow
any #include statements. Fix this by installing a copy in $prefix/libdir
and patching the build system to look for the file there.

Python's build system also reads a copy of a generated Makefile that it
installs in $prefix/libdir and uses some of its variables in module
builds. This causes contamination because the include path points to
a recipe specific sysroot. Clear the path from this Makefile with a sed
expression and patch the module build system to not pass a -L'' flag if
the variable is empty. The path is not needed, because it is in the
default search path of the cross compiler toolchain.

Signed-off-by: Clemens Lang <clemens.lang@bmw-carit.de>